AstarのTarget位置更新追加・StageModel原点を変更・Wall判定を少し正確に(スピード早いとめり込む)

+++ CollisionMap.h
@@ -40,7 +40,7 @@ public:
     void CalcMapWall(XMFLOAT3& _pos, float speed);
 
     //CollisionMap‚Æ‚Ì”»’èŽžŠÔ“à‚©‚ç‚Ì‚â‚Â
-    bool IsWall(float x, float z);
+    bool IsWall(int x, int z);
 
 };
